+++ adm/processa/proc_apagar_produto.php
@@ -11,8 +11,28 @@ if(isset($_SESSION['usuarioNome'])){
 
 include_once("../conexao.php");
 $id 				= $_GET["id"];
+$ativ_inativ		= 0;
 
-$query = "DELETE FROM produtos WHERE id=$id";
+if(isset($_GET["inativ"])){$ativ_inativ = $_GET["inativ"];}
+$ativ = 0;
+$msg = "";
+
+if ($ativ_inativ > 0) {
+	$pre_query = "SELECT id, situacao_id FROM produtos WHERE id=$id";
+	$pre_resultado = mysqli_query($conectar,$pre_query);
+	
+	while($pre_linhas = mysqli_fetch_array($pre_resultado)){
+		if ($pre_linhas['situacao_id'] > 1 ){ $ativ = 1; $msg = "ativado"; } else { $ativ = 2; $msg = "desativado"; }
+	}
+
+	$query = "UPDATE produtos set situacao_id = '$ativ' WHERE id=$id";
+
+} else { 
+	$query = "DELETE FROM produtos WHERE id=$id";
+	$msg = "apagado";
+}
+
+
 $resultado = mysqli_query($conectar,$query);
 $linhas = mysqli_affected_rows($conectar);
 
@@ -29,7 +49,7 @@ $linhas = mysqli_affected_rows($conectar);
 			echo "
 				<META HTTP-EQUIV=REFRESH CONTENT = '0;URL=http://".$_SERVER['HTTP_HOST']."/adm/administrativo.php?link=10'>
 				<script type=\"text/javascript\">
-					alert(\"Produto apagado com Sucesso.\");
+					alert(\"Produto ".$msg." com sucesso.\");
 				</script>
 			";		   
 		}
@@ -37,7 +57,7 @@ $linhas = mysqli_affected_rows($conectar);
 				echo "
 				<META HTTP-EQUIV=REFRESH CONTENT = '0;URL=http://".$_SERVER['HTTP_HOST']."/adm/administrativo.php?link=10'>
 				<script type=\"text/javascript\">
-					alert(\"Produto não foi apagado com Sucesso.\");
+					alert(\"Produto não foi ".$msg." com sucesso.\");
 				</script>
 			";
